move uniform angle, uniform solid angle into angular; add __all__ to all files
